+++ tests/gtest/avifpng16bittest.cc
@@ -32,16 +32,17 @@ ImagePtr ReadImageLosslessBitDepth(const std::string& path,
   image->matrixCoefficients = AVIF_MATRIX_COEFFICIENTS_IDENTITY;
 
   uint32_t output_depth;
-  if (avifReadImage(path.c_str(),
-                    AVIF_APP_FILE_FORMAT_UNKNOWN /* guess format */,
-                    AVIF_PIXEL_FORMAT_YUV444, bit_depth,
-                    AVIF_CHROMA_DOWNSAMPLING_AUTOMATIC,
-                    /*ignoreColorProfile=*/true,
-                    /*ignoreExif=*/true, /*ignoreXMP=*/true,
-                    /*ignoreGainMap=*/true, AVIF_DEFAULT_IMAGE_SIZE_LIMIT,
-                    image.get(), &output_depth,
-                    /*sourceTiming=*/nullptr,
-                    /*frameIter=*/nullptr) == AVIF_APP_FILE_FORMAT_UNKNOWN) {
+  if (avifReadImage(
+          path.c_str(), AVIF_APP_FILE_FORMAT_UNKNOWN /* guess format */,
+          AVIF_PIXEL_FORMAT_YUV444, bit_depth,
+          AVIF_CHROMA_DOWNSAMPLING_AUTOMATIC,
+          /*requestedFormatGainMap=*/AVIF_PIXEL_FORMAT_NONE /* auto */,
+          /*ignoreColorProfile=*/true,
+          /*ignoreExif=*/true, /*ignoreXMP=*/true,
+          /*ignoreGainMap=*/true, AVIF_DEFAULT_IMAGE_SIZE_LIMIT, image.get(),
+          &output_depth,
+          /*sourceTiming=*/nullptr,
+          /*frameIter=*/nullptr) == AVIF_APP_FILE_FORMAT_UNKNOWN) {
     return nullptr;
   }
   EXPECT_EQ(output_depth, bit_depth);